+++ rusoto/services/comprehendmedical/src/lib.rs
@@ -1,0 +1,32 @@
+
+// =================================================================
+//
+//                           * WARNING *
+//
+//                    This file is generated!
+//
+//  Changes made to this file will be overwritten. If changes are
+//  required to the generated code, the service_crategen project
+//  must be updated to generate the changes.
+//
+// =================================================================
+
+#![doc(html_logo_url = "https://raw.githubusercontent.com/rusoto/rusoto/master/assets/logo-square.png")]
+//! <p> Comprehend Medical extracts structured information from unstructured clinical text. Use these actions to gain insight in your documents. </p>
+//!
+//! If you're using the service, you're probably looking for [ComprehendMedicalClient](struct.ComprehendMedicalClient.html) and [ComprehendMedical](trait.ComprehendMedical.html).
+
+extern crate bytes;
+extern crate futures;
+extern crate rusoto_core;
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate serde_json;
+
+mod generated;
+mod custom;
+
+pub use generated::*;
+pub use custom::*;
+