+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add Kafka service
 - Add Chime service
 - Add RDS Data service
+- Add ComprehendMedical service
 
 ## [0.38.0] - 2019-04-17